+++ exodus/urls.py
@@ -24,6 +24,7 @@ urlpatterns = [
     path("accounts/", include("accounts.urls")),
     path("", include("news.urls")),
     path("", include("starmap.urls")),
+    path("", include("starships.urls")),
     path("", include("comms.urls")),
     path("", include("characters.urls")),
     path("", include("agencies.urls")),